Refactored AboutCard buttons into individual components
