+++ plugins/catalog/src/components/AboutCard/AboutCard.tsx
@@ -15,46 +15,23 @@
  */
 
 import {
-  ANNOTATION_EDIT_URL,
-  ANNOTATION_LOCATION,
-  DEFAULT_NAMESPACE,
-  stringifyEntityRef,
-} from '@backstage/catalog-model';
-import {
   HeaderIconLinkRow,
-  IconLinkVerticalProps,
   InfoCardVariants,
-  Link,
 } from '@backstage/core-components';
-import {
-  alertApiRef,
-  errorApiRef,
-  useApi,
-  useRouteRef,
-} from '@backstage/core-plugin-api';
-import {
-  ScmIntegrationIcon,
-  scmIntegrationsApiRef,
-} from '@backstage/integration-react';
-import {
-  catalogApiRef,
-  getEntitySourceLocation,
-  useEntity,
-} from '@backstage/plugin-catalog-react';
+import { useEntity } from '@backstage/plugin-catalog-react';
 import {
   Card,
   CardContent,
   CardHeader,
   Divider,
-  IconButton,
   makeStyles,
 } from '@material-ui/core';
-import CachedIcon from '@material-ui/icons/Cached';
-import DocsIcon from '@material-ui/icons/Description';
-import EditIcon from '@material-ui/icons/Edit';
-import React, { useCallback } from 'react';
-import { viewTechDocRouteRef } from '../../routes';
+import React from 'react';
 import { AboutContent } from './AboutContent';
+import { EditMetadataButton } from './EditMetadataButton';
+import { RefreshActionButton } from './RefreshActionButton';
+import { ViewInSourceButton } from './ViewInSourceButton';
+import { ViewInTechDocsButton } from './ViewInTechDocsButton';
 
 const useStyles = makeStyles({
   gridItemCard: {
@@ -92,39 +69,6 @@ export function AboutCard(props: AboutCardProps) {
   const { variant } = props;
   const classes = useStyles();
   const { entity } = useEntity();
-  const scmIntegrationsApi = useApi(scmIntegrationsApiRef);
-  const catalogApi = useApi(catalogApiRef);
-  const alertApi = useApi(alertApiRef);
-  const errorApi = useApi(errorApiRef);
-  const viewTechdocLink = useRouteRef(viewTechDocRouteRef);
-
-  const entitySourceLocation = getEntitySourceLocation(
-    entity,
-    scmIntegrationsApi,
-  );
-  const entityMetadataEditUrl =
-    entity.metadata.annotations?.[ANNOTATION_EDIT_URL];
-
-  const viewInSource: IconLinkVerticalProps = {
-    label: 'View Source',
-    disabled: !entitySourceLocation,
-    icon: <ScmIntegrationIcon type={entitySourceLocation?.integrationType} />,
-    href: entitySourceLocation?.locationTargetUrl,
-  };
-  const viewInTechDocs: IconLinkVerticalProps = {
-    label: 'View TechDocs',
-    disabled:
-      !entity.metadata.annotations?.['backstage.io/techdocs-ref'] ||
-      !viewTechdocLink,
-    icon: <DocsIcon />,
-    href:
-      viewTechdocLink &&
-      viewTechdocLink({
-        namespace: entity.metadata.namespace || DEFAULT_NAMESPACE,
-        kind: entity.kind,
-        name: entity.metadata.name,
-      }),
-  };
 
   let cardClass = '';
   if (variant === 'gridItem') {
@@ -140,46 +84,21 @@ export function AboutCard(props: AboutCardProps) {
     cardContentClass = classes.fullHeightCardContent;
   }
 
-  const entityLocation = entity.metadata.annotations?.[ANNOTATION_LOCATION];
-  // Limiting the ability to manually refresh to the less expensive locations
-  const allowRefresh =
-    entityLocation?.startsWith('url:') || entityLocation?.startsWith('file:');
-  const refreshEntity = useCallback(async () => {
-    try {
-      await catalogApi.refreshEntity(stringifyEntityRef(entity));
-      alertApi.post({ message: 'Refresh scheduled', severity: 'info' });
-    } catch (e) {
-      errorApi.post(e);
-    }
-  }, [catalogApi, alertApi, errorApi, entity]);
-
   return (
     <Card className={cardClass}>
       <CardHeader
         title="About"
         action={
           <>
-            {allowRefresh && (
-              <IconButton
-                aria-label="Refresh"
-                title="Schedule entity refresh"
-                onClick={refreshEntity}
-              >
-                <CachedIcon />
-              </IconButton>
-            )}
-            <IconButton
-              component={Link}
-              aria-label="Edit"
-              disabled={!entityMetadataEditUrl}
-              title="Edit Metadata"
-              to={entityMetadataEditUrl ?? '#'}
-            >
-              <EditIcon />
-            </IconButton>
+            <RefreshActionButton />
+            <EditMetadataButton />
           </>
         }
-        subheader={<HeaderIconLinkRow links={[viewInSource, viewInTechDocs]} />}
+        subheader={
+          <HeaderIconLinkRow
+            links={[<ViewInSourceButton />, <ViewInTechDocsButton />]}
+          />
+        }
       />
       <Divider />
       <CardContent className={cardContentClass}>
